[Tests] Ensure structured template works when no root keys
I encountered an issue when testing the `-t structured` option on one of my own projects Strings files. The problem was that because the template was checking for `structuredStrings.strings` a few commits ago it would produce the empty tamplate part even when many keys with dot-level higher than 1 were defined (like in my project). This was fixed in 5aec504abf8aba42b08129750e68a087759ded03 by checking for the existens of the structuredStrings only.