Resolve binary concrete path if it is a symbolic link
Seems like there is an issue if we call swiftgen using a symbolic link.
The issue does not come from swiftgen implementation but seems to be an
Apple bug which does not return the same `arguments[0]` depending on
where you call the symbolic link from.
[This sample](https://github.com/Liquidsoul/apple-radars/tree/master/symlink-arg0)
shows the issue.
In my opinion, the value should always be the symbolic link path,
meaning that bundled templates resolution would not work. So, in order
to solve this, swiftgen needs to go to the concrete swiftgen path.