More secure TestHelper using ContextConvertible
The small drawback of this is that we end up with `public protocol ContextConvertible` in SwiftGenKit framework which is in practice used as a marker just for tests.
But at least that avoids any test _crash_ in case a case goes wrong, and avoids the risk of false positives in tests by restricting the tests to try to convert to Dict only when allowed.

- Without this, there would be a risk that 2 objects of unhandled class (be it a public class or some internal Swift object like `__StringStorage` or similar) would end up being tried to be converted into NSDictionary, and if those unhandled classes have no `Mirror.Children` with public labels, that will make the test crash instead of failing.

 - With this marker protocol, we ensure that only the types we have marked as `ContextConvertible` will be decoded to dictionaries by the tests