fix: rake output:compile for XCAssets templates by removing lazy vars
The Swift compiler does not allow the use of @available on stored variables
anymore. This means that the `lazy var` used in the templates cannot be made
available only for certain OSes.
I have fixed the issue by removing the `lazy var` altogether. Compared to the
previous implementation, this introduces an overhead on every access to a
color. I could not find a way to keep the lazy var without droping old OSes.